// web-front/src/types/typescript-axios/api.ts
export interface TodoEntity {
    /**
     * 
     * @type {number}
     * @memberof TodoEntity
     */
    'id': number;
    /**
     * 
     * @type {string}
     * @memberof TodoEntity
     */
    'title': string;
    /**
     * 
     * @type {string}
     * @memberof TodoEntity
     */
    'content': string;
    /**
     * 
     * @type {string}
     * @memberof TodoEntity
     */
    'userId': string;
    /**
     * 
     * @type {string}
     * @memberof TodoEntity
     */
    'createdAt': string;
    /**
     * 
     * @type {string}
     * @memberof TodoEntity
     */
    'updatedAt': string;
}